pyproject.toml
[tool.ruff]
select = ['E', 'F', 'Q']
ignore = ['W191', 'E111', 'E114', 'E117', 'D206', 'D300', 'Q000', 'Q001', 'Q002', 'Q003', 'COM812', 'COM819', 'ISC001', 'ISC002']
exclude = ['conftest.py', 'dev.py', 'test.py']
line-length = 120

[tool.ruff.format]
quote-style = 'single'
preview = true

[tool.ruff.pydocstyle]
convention = 'google'

[tool.poetry]
name = "snack"
version = "0.1.0"
description = "Snack management system"
authors = ["Jongtaek Hwang <orchwang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.9"
django-filter = "^23.5"
psycopg2-binary = "^2.9.9"
Markdown = "^3.5.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.0"
requests = "^2.31.0"
shortuuid = "^1.0.11"
pillow = "^10.2.0"
django-cors-headers = "^4.3.1"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
redis = "^5.0.1"
django-redis = "^5.4.0"

[tool.poetry.group.dev.dependencies]
poetry = "^1.7.1"
pytest = "^7.4.4"
pytest-django = "^4.7.0"
pytest-xdist = "^3.5.0"
coverage = "^7.4.0"
requests-mock = "^1.11.0"
pytest-celery = "^0.0.0"
fakeredis = "^2.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
